refactor homepage components for better re-use and modularity
