+++ src/ui/src/components/ui/data-table.tsx
@@ -1,0 +1,67 @@
+import {
+  type ColumnDef,
+  flexRender,
+  type Table as TanstackTable,
+} from "@tanstack/react-table";
+import {
+  Table,
+  TableBody,
+  TableCell,
+  TableHead,
+  TableHeader,
+  TableRow,
+} from "./table";
+
+interface DataTableProps<TData> {
+  table: TanstackTable<TData>;
+  columns: ColumnDef<TData, unknown>[];
+  emptyMessage?: string;
+  className?: string;
+}
+
+export function DataTable<TData>({
+  table,
+  columns,
+  emptyMessage = "No results.",
+  className,
+}: DataTableProps<TData>) {
+  return (
+    <Table className={className}>
+      <TableHeader>
+        {table.getHeaderGroups().map((headerGroup) => (
+          <TableRow key={headerGroup.id} className="hover:bg-transparent">
+            {headerGroup.headers.map((header) => (
+              <TableHead
+                key={header.id}
+                style={{ width: header.getSize() !== 150 ? header.getSize() : undefined }}
+              >
+                {header.isPlaceholder
+                  ? null
+                  : flexRender(header.column.columnDef.header, header.getContext())}
+              </TableHead>
+            ))}
+          </TableRow>
+        ))}
+      </TableHeader>
+      <TableBody>
+        {table.getRowModel().rows.length === 0 ? (
+          <TableRow className="hover:bg-transparent">
+            <TableCell colSpan={columns.length} className="text-center text-muted-foreground py-8">
+              {emptyMessage}
+            </TableCell>
+          </TableRow>
+        ) : (
+          table.getRowModel().rows.map((row) => (
+            <TableRow key={row.id}>
+              {row.getVisibleCells().map((cell) => (
+                <TableCell key={cell.id}>
+                  {flexRender(cell.column.columnDef.cell, cell.getContext())}
+                </TableCell>
+              ))}
+            </TableRow>
+          ))
+        )}
+      </TableBody>
+    </Table>
+  );
+}
